requirements.txt
essentia-tensorflow==2.1b6.dev1389
numpy>=1.21.0
pandas>=1.3.0
urllib3>=1.26.0